requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# LangChain and AI/ML
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8
unstructured==0.11.8

# Text processing & NLP
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.1
numpy==1.24.3
pandas==2.1.4

# Vector database
faiss-cpu==1.7.4
chromadb==0.4.18

# AI/ML
openai==1.3.7
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
tiktoken==0.5.2

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0

#frontend
streamlit
requests